rss_stat: Add support to uniquely identify mm
When a process updates the RSS of a different process, the rss_stat
tracepoint appears in the context of the process doing the update. This
can confuse userspace that the RSS of process doing the update is
updated, while in reality a different process's RSS was updated.

This can happen from the following paths:
1. direct reclaim
2. background reclaim (kswapd)
3. userspace initiated reclaim which Minchan has been working on.

This patch adds more information to the tracepoint about the mm being
updated. When the tracepoint fires, the patch hashes the mm pointer into
a unique ID and stores it within the tracepoint. Userspace can then
uniquely identify the process being updated based on the mm's hash.

Reported-by: Ioannis Ilkos <ilkos@google.com>
Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>

diff-note-start
This is an internal RFC patch for Bug: 140711541. I'd like any feedback
before I can post upstream, thanks.

option-prefix RFC
option-start
Cc: keescook@chromium.org
Cc: minchan@google.com
Cc: timmurray@google.com
Cc: primiano@google.com
Cc: ilkos@google.com
Cc: carmenjackson@google.com
Cc: fmayer@google.com
Cc: hjd@google.com
Cc: ilkos@google.com
Cc: joaodias@google.com
Cc: joelaf@google.com
Cc: keescook@google.com
Cc: lalitm@google.com
Cc: minchan@google.com
Cc: primiano@google.com
Cc: rslawik@google.com
Cc: sspatil@google.com
Cc: timmurray@google.com