Built-in filters for trace events
These patches allow a user to define a configurable trace event.

USECASE
-------
The rss_stat tracepoint is very noisy. We want this tracepoint to be
emitted to traces on very specific conditions. For our usecase, we want
it emitted whenever the signal crosses a custom boundary, such as 512KB.
Another user may want to emit this for every page (4KB), it is really
custom and user-dependent.

It is not possible to specify such cases using ftrace's filters. Further
eBPF programs for this would be overkill as that forces the user to use
the perf_events subsystem.

SOLUTION
--------
Define a new TRACE_EVENT macro called TRACE_EVENT_FILTERED, and related
macros to accept a callback. The callback is a function pointer that is
called to be able to do the filtering. If the function returns true,
then ftrace filters the trace event.

In order for the function to do its job, the user can specify the
configuration they need (for example, rss_stat boundary in the above
example). The callback will be passed this configuration. The user uses
new tracefs entries to set the parameters for the filtering. The tracefs
entries are automatically created in tracing/events/<subsys>/<event>/conf/*
Currently the patches only support integer values but others could be
added.

TODO
----
1. Combine the functionality of DEFINE_BUILTIN_FILTER with
   TRACE_EVENT_FILTERED and eliminate DEFINE_BUILTIN_FILTER.
2. Add custom types to the configuration structure. Currently only
   integers are supported.
3. Add an example of a rate-limiting filter.
4. Cleanup and post RFC.

Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>