Clean up trace_events.h
These patches clean up the trace_events.h header which is a mess. We
split the header into a new _stage.h which contains all the undefines
for every stage. This header is included every stage so that all the
macro undefines happen before every stage so and a stage starts clean.

Motivation: Recently I ran into an issue where I added a new stage but
found that a previous stage defined some macros, this took a lot of time
to debug. I would like to clean this process up.

Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>