Pause sibling if core is in priv state
Corner case:
This should also handle MDS due to flushing of the unpartitioned
arch buffers before entering into C-state.
So it does handle the below scenario:
1. HT1 running attacker
2. HT2 gets IRQ and sends IPI
3. HT1 busy loops in IPI
4. HT2 finishes IRQ and resets flag
5. HT1 returns to usermode and starts executing attacker.
6. HT2 transitions to idle and unpartitions the buffers, thus sharing
   the store buffer with HT1.

One question might be, if there is a huge delay between step 5 and step
6. Can HT1 see HT2's store buffer? The answer is no. Because HT2 has
still not entered C-state, so the store buffer remains partitioned.
And, HT2 clears the unpartitioned buffers before it transitions to idle
C-state, due to MDS mitigations. This is documentered here in 16.5.2:
https://www.kernel.org/doc/html/latest/x86/mds.html

Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>
Change-Id: I045b32feb601a61fbb56d4d42ee02a36e45f76d2