Don't resend IPI if a previous one is in progress
Sending an async smp call, while a previous one was not received (or CSD
was not yet unlocked) seems to trigger warnings and cause problems. This
simple mechanism fixes it. If a 'pause IPI' is pending, just don't send
another one.

In the future, we can implement logic where if a 'pause IPI' is pending,
we wait for it to finish.

Another way to implement this is just wait for CSD to get unlocked.
However we don't have APIs for this yet, and it is unclear if that
API will be accepted upstream.

Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>
Change-Id: I327dfebefb6b1f513d0a8fa06035e7b2711c700b